Seq. Omniglot for classification

+++ SiameseNetworks/model.py
@@ -571,7 +571,7 @@ import numpy as np
 def _prune_freeze(module, task):
     if any([isinstance(module, ALinear), isinstance(module, AConv2d)]):
         mask = (module.soft_round(module.adjx[task]) <= 0.15).data
-        print(mask.sum()/np.prod(mask.shape))
+        print(mask.sum().float()/np.prod(mask.shape))
         for k in range(len(module.adjx)):
             if k==task:
                 continue
